ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.0
python-dotenv==1.0.0